# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "见微 API",
    "version": "1.0.0",
    "description": "Anonymous device API for privacy-filtered photo analysis and reviewed knowledge cards."
  },
  "servers": [{ "url": "https://api.example.invalid" }],
  "paths": {
...
    "/v1/devices/register": {
      "post": {
        "operationId": "registerDevice",
        "x-android-client": "retrofit",
        "x-android-request-model": "RegisterRequest",
        "x-android-response-model": "RegisterResponse",
        "x-backend-zod-schema": "registerDeviceSchema",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } } } },
        "responses": { "201": { "description": "Registered", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterResponse" } } } }, "400": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/v1/analysis-jobs": {
      "post": {
        "operationId": "createAnalysisJob",
        "x-android-client": "retrofit",
        "x-android-request-model": "CreateJobRequest",
        "x-android-response-model": "CreateJobResponse",
        "x-backend-zod-schema": "createAnalysisJobSchema",
        "security": [{ "deviceBearer": [] }],
        "parameters": [{ "$ref": "#/components/parameters/EvaluationLease" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateJobRequest" } } } },
        "responses": { "201": { "description": "Created or reused", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateJobResponse" } } } }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "410": { "$ref": "#/components/responses/Error" }, "429": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/v1/analysis-jobs/{id}/image": {
      "put": {
        "operationId": "uploadAnalysisImage",
        "x-android-client": "raw",
        "security": [{ "deviceBearer": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "requestBody": { "required": true, "content": { "image/jpeg": { "schema": { "type": "string", "contentEncoding": "binary" } } } },
        "responses": { "200": { "description": "Uploaded", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UploadJobResponse" } } } }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "410": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/v1/analysis-jobs/{id}/complete": {
      "post": {
        "operationId": "completeAnalysisJob",
        "x-android-client": "retrofit",
        "x-android-response-model": "CompleteJobResponse",
        "x-backend-zod-schema": "completeAnalysisJobSchema",
        "security": [{ "deviceBearer": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "requestBody": { "required": false, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CompleteJobRequest" } } } },
        "responses": { "200": { "description": "Completed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CompleteJobResponse" } } } }, "401": { "$ref": "#/components/responses/Error" }, "402": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/v1/analysis-jobs/{id}": {
      "get": {
        "operationId": "getAnalysisJob",
        "security": [{ "deviceBearer": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": { "200": { "description": "Job status", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobStatusResponse" } } } }, "401": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/v1/cards": {
      "get": {
        "operationId": "listCards",
        "x-android-client": "retrofit",
        "x-android-response-model": "CardsResponse",
        "security": [{ "deviceBearer": [] }],
        "parameters": [
          { "name": "cursor", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 50, "default": 20 } }
        ],
        "responses": { "200": { "description": "Cards", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CardsResponse" } } } }, "401": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/v1/cards/select-daily": {
      "post": {
        "operationId": "selectDailyCard",
        "x-backend-zod-schema": "selectDailyCardSchema",
        "security": [{ "deviceBearer": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SelectDailyCardRequest" } } } },
        "responses": { "200": { "description": "AI-selected daily card", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DailyCardSelectionResponse" } } } }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "402": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "502": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/v1/cards/{id}/feedback": {
      "post": {
        "operationId": "sendFeedback",
        "x-android-client": "retrofit",
        "x-android-request-model": "FeedbackRequest",
        "x-android-response-model": "FeedbackResponse",
        "x-backend-zod-schema": "feedbackSchema",
        "security": [{ "deviceBearer": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FeedbackRequest" } } } },
        "responses": { "201": { "description": "Recorded", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FeedbackResponse" } } } }, "401": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/v1/items/{cardId}/track": {
      "post": {
        "operationId": "trackItem",
        "x-android-client": "retrofit",
        "x-android-request-model": "TrackRequest",
        "x-android-response-model": "TrackItemResponse",
        "x-backend-zod-schema": "trackItemSchema",
        "security": [{ "deviceBearer": [] }],
        "parameters": [{ "name": "cardId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TrackRequest" } } } },
        "responses": { "201": { "description": "Tracked", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TrackItemResponse" } } } }, "401": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } }
      },
      "delete": {
        "operationId": "cancelTrackedItem",
        "x-android-client": "retrofit",
        "x-android-response-model": "UntrackItemResponse",
        "security": [{ "deviceBearer": [] }],
        "parameters": [{ "name": "cardId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Tracking cancelled or already absent", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UntrackItemResponse" } } } }, "401": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/v1/device-data": {
      "delete": {
        "operationId": "deleteDeviceData",
        "x-android-client": "retrofit",
        "x-android-response-model": "DeleteDeviceDataResponse",
        "security": [{ "deviceBearer": [] }],
        "responses": { "200": { "description": "Deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeleteDeviceDataResponse" } } } }, "401": { "$ref": "#/components/responses/Error" } }
      }
    }
  },
  "components": {
    "securitySchemes": { "deviceBearer": { "type": "http", "scheme": "bearer" } },
    "parameters": {
      "Id": { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
      "EvaluationLease": {
        "name": "X-Jianwei-Evaluation-Lease",
        "in": "header",
        "required": false,
        "description": "Short-lived bounded bearer for the Debug-only authorized evaluation runner; must be supplied with evaluationContext.",
        "schema": { "type": "string", "pattern": "^[A-Za-z0-9_-]{43}$" }
      }
    },
    "responses": { "Error": { "description": "Structured API error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } } },
    "schemas": {
      "RegisterRequest": { "type": "object", "additionalProperties": false, "required": ["installationId"], "properties": { "installationId": { "type": "string", "format": "uuid" } } },
      "RegisterResponse": { "type": "object", "additionalProperties": false, "required": ["deviceId", "deviceToken", "installationBindingSha256", "created"], "properties": { "deviceId": { "type": "string", "format": "uuid" }, "deviceToken": { "type": "string", "minLength": 43, "maxLength": 43, "pattern": "^[A-Za-z0-9_-]{43}$" }, "installationBindingSha256": { "type": "string", "pattern": "^[a-f0-9]{64}$", "description": "Domain-separated SHA-256 binding to the submitted installationId; the raw installation identifier is never echoed." }, "created": { "type": "boolean" } } },
      "CreateJobRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["candidateToken", "localLabels", "qualityScore", "sensitiveFlags", "contentType"],
        "properties": {
          "candidateToken": { "type": "string", "format": "uuid" },
          "capturedAtBucket": { "type": ["string", "null"], "format": "date" },
          "localLabels": { "type": "array", "maxItems": 20, "items": { "type": "string", "maxLength": 80 } },
          "qualityScore": { "type": "number", "minimum": 0, "maximum": 1 },
          "sensitiveFlags": { "type": "array", "maxItems": 20, "items": { "type": "string", "maxLength": 80 } },
          "contentType": { "const": "image/jpeg" },
          "evaluationContext": { "$ref": "#/components/schemas/EvaluationContextRequest" }
        }
      },
...
      "CreateJobResponse": { "type": "object", "additionalProperties": false, "required": ["jobId", "candidateToken", "status", "uploadUrl", "uploadSessionId", "expiresAt"], "properties": { "jobId": { "type": "string", "format": "uuid" }, "candidateToken": { "type": "string", "format": "uuid" }, "status": { "type": "string", "enum": ["awaiting_upload", "uploaded", "completed", "needs_content", "rejected"] }, "uploadUrl": { "oneOf": [{ "type": "string", "format": "uri" }, { "type": "string", "maxLength": 0 }, { "type": "null" }] }, "uploadSessionId": { "type": ["string", "null"], "format": "uuid" }, "expiresAt": { "type": "string", "format": "date-time" } } },
      "UploadJobResponse": { "type": "object", "additionalProperties": false, "required": ["jobId", "candidateToken", "uploadSessionId", "status"], "properties": { "jobId": { "type": "string", "format": "uuid" }, "candidateToken": { "type": "string", "format": "uuid" }, "uploadSessionId": { "type": "string", "format": "uuid" }, "status": { "type": "string", "const": "uploaded" } } },
...
      "CompleteJobRequest": { "type": "object", "additionalProperties": false, "properties": { "modelAccess": { "$ref": "#/components/schemas/ModelAccess" } } },
...
      "CardsResponse": { "type": "object", "additionalProperties": false, "required": ["items", "nextCursor"], "properties": { "items": { "type": "array", "maxItems": 50, "items": { "$ref": "#/components/schemas/Card" } }, "nextCursor": { "type": ["string", "null"], "format": "uuid" } } },
...
    }
  }
}